pyproject.toml
[project]
name = "nucwave2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "biopython>=1.86",
    "numpy>=2.4.1",
    "polars>=1.37.1",
    "rich>=14.2.0",
    "typer>=0.21.1",
]
